# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Invoice Management
  version: 0.0.1
servers:
- url: /
...
paths:
...
  /api/v{version}/customers:
    get:
      tags:
      - Customer
      operationId: api_vversion_customers_get
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: name
        in: query
        description: Customer's name
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: cnpj
        in: query
        description: Customer's cnpj
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Customer founded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCustomers'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.customer_controller
    post:
      tags:
      - Customer
      operationId: api_vversion_customers_post
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Register customer
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        "201":
          description: Customer successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.customer_controller
  /api/v{version}/customers/{customerID}:
    get:
      tags:
      - Customer
      operationId: api_vversion_customers_customer_idget
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: customerID
        in: path
        description: Customer's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Customer founded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCustomer'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.customer_controller
    put:
      tags:
      - Customer
      operationId: api_vversion_customers_customer_idput
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: customerID
        in: path
        description: Customer's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        description: Update customer
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.customer_controller
  /api/v{version}/customers/{customerID}/archives:
    put:
      tags:
      - Customer
      operationId: api_vversion_customers_customer_id_archives_put
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: customerID
        in: path
        description: Customer's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.customer_controller
...
  /api/v{version}/expenses/{categoryID}:
    post:
      tags:
      - Expense
      operationId: api_vversion_expenses_category_idpost
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: categoryID
        in: path
        description: Category's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        description: Expense by category
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Expense'
      responses:
        "201":
          description: Expense's ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.expense_controller
  /api/v{version}/expenses/{expenseID}:
    put:
      tags:
      - Expense
      operationId: api_vversion_expenses_expense_idput
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: expenseID
        in: path
        description: Expense's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        description: Register expense
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Expense'
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.expense_controller
    delete:
      tags:
      - Expense
      operationId: api_vversion_expenses_expense_iddelete
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: expenseID
        in: path
        description: Expense's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.expense_controller
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        cnpj:
          type: string
        commercial_name:
          type: string
        legal_name:
          type: string
      example:
        commercial_name: commercial_name
        cnpj: cnpj
        legal_name: legal_name
    CustomerResponse:
      type: object
      properties:
        customer_id:
          type: integer
          format: int32
      example:
        customer_id: 0
    GetCustomer:
      type: object
      properties:
        customer:
          $ref: '#/components/schemas/Customer'
      example:
        customer:
          commercial_name: commercial_name
          cnpj: cnpj
          legal_name: legal_name
    GetCustomers:
      type: object
      properties:
        count:
          type: integer
          format: int32
        customers:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
      example:
        count: 0
        customers:
        - commercial_name: commercial_name
          cnpj: cnpj
          legal_name: legal_name
        - commercial_name: commercial_name
          cnpj: cnpj
          legal_name: legal_name
...
    Expense:
      type: object
      properties:
        amount:
          type: number
        description:
          type: string
        accrual_date:
          type: string
        transaction_date:
          type: string
        customer_id:
          type: integer
    ExpenseResponse:
      type: object
      properties:
        expense_id:
          type: integer
      example:
        expense_id: 0
...
    Error:
      type: object
      properties:
        error:
          type: string
  securitySchemes:
    user_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_user_auth